↗️ Improved compatibilty of build process: Using docker buildx
By using docker buildx, we are able to build the Docker image for ARM based system when the python base image is in play. This was not possible before.
This leads into better compatibility with newer python features and a way easier upgrade process since we are able to use the pre build python base instead of compiling the wished versions by ourselves.