Rewrite channel listener / creator use new database
This is nearly a complete rewrite.
It was developed in a new file and copy-pasted into this one when it was done.

Does NOT include the setup-voice command so far.

I copied it over since git detects some similarities which make the diff more interesting than a "delete one file, add an other"